+++ packages/antd-management-fast-design-playground/src/components/displayer/DocumentPrintDesigner/DocumentContent/GeneralConfigs/GeneralExtraConfigBoxContainer/index.jsx
@@ -1,0 +1,14 @@
+import { Divider } from 'antd';
+import React from 'react';
+
+function GeneralExtraConfigBoxContainer({ showDivider = true, children }) {
+  return (
+    <div>
+      {children}
+
+      {showDivider ? <Divider style={{ margin: '6px 0' }} /> : null}
+    </div>
+  );
+}
+
+export { GeneralExtraConfigBoxContainer };