+++ packages/antd-management-fast-design-playground/src/components/displayer/DocumentPrintDesigner/DocumentContent/GeneralConfigs/GeneralGridColor/index.jsx
@@ -21,7 +21,7 @@ function GeneralGridColor(properties) {
       <GeneralConfigContainer>
         <FlexBox
           flexAuto="left"
-          left={<VerticalBox>边框颜色：</VerticalBox>}
+          left={<VerticalBox>边框色：</VerticalBox>}
           right={
             <ColorPicker
               value={gridColor}