+++ packages/antd-management-fast-design-playground/src/components/displayer/DocumentPrintDesigner/DocumentContent/TitleConfigBox/index.jsx
@@ -149,7 +149,7 @@ function TitleConfigBox(properties) {
         />
       </TitleConfigBoxContainer>
 
-      <TitleConfigBoxContainer>
+      <TitleConfigBoxContainer showDivider={false}>
         <FlexBox
           flexAuto="left"
           left={<VerticalBox>字体大小：</VerticalBox>}